cleanup: extract test helpers and remove duplicate test_watchlist.py (#341)
- Delete test_watchlist.py, move its batch endpoint tests into test_groups.py
- Extract _create_asset from 4 repo test files into helpers.create_test_asset
- Extract _make_asset from service tests into helpers.make_model_asset and
  helpers.make_mock_asset (two patterns: real model vs MagicMock)
- Remove dead _mock_validate function from test_assets.py
- Remove unused Group/group_assets imports from test_asset_repo.py
