+++ backend/tests/repositories/test_thesis_repo.py
@@ -4,18 +4,9 @@ import pytest
 
 from app.models import Asset, AssetType, Thesis
 from app.repositories.thesis_repo import ThesisRepository
+from tests.helpers import create_test_asset as _create_asset
 
 pytestmark = pytest.mark.asyncio(loop_scope="function")
-
-
-async def _create_asset(db, symbol: str, **kwargs) -> Asset:
-    defaults = dict(name=f"{symbol} Inc.", type=AssetType.STOCK, currency="USD")
-    defaults.update(kwargs)
-    asset = Asset(symbol=symbol, **defaults)
-    db.add(asset)
-    await db.commit()
-    await db.refresh(asset)
-    return asset
 
 
 async def test_get_by_asset_returns_none_when_not_found(db):
